backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Wiper-R <rshivang12345@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
quart = "^0.16.2"
tortoise-orm = "<=0.18.0"
aerich = "^0.6.1"
pydantic = "^1.8.2"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
aiohttp = "^3.8.1"
Quart-CORS = "^0.5.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}


[tool.aerich]
tortoise_orm = "config.TORTOISE"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
